有无符号混用告警整改

+++ frameworks/innerkitsimpl/medialibrary_data_extension/include/medialibrary_album_operations.h
@@ -44,7 +44,7 @@ struct MergeAlbumInfo {
     int rank;
     int renameOperation;
     std::string albumName;
-    int isCoverSatisfied;
+    uint8_t isCoverSatisfied;
     std::string tagId;
 };
 class MediaLibraryAlbumOperations {
